ADDED :Token and change the return type

+++ src/main/java/com/bridgelabz/lmscandidateservice/service/CandidateService.java
@@ -4,9 +4,11 @@ import com.bridgelabz.lmscandidateservice.dto.CandidateDTO;
 import com.bridgelabz.lmscandidateservice.exception.LMSException;
 import com.bridgelabz.lmscandidateservice.model.CandidateModel;
 import com.bridgelabz.lmscandidateservice.repository.CandidateRepository;
+import com.bridgelabz.lmscandidateservice.util.Response;
 import com.bridgelabz.lmscandidateservice.util.TokenUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -20,10 +22,8 @@ public class CandidateService implements ICandidateService {
     MailService mailService;
     @Autowired
     CandidateRepository candidateRepository;
-//    @Autowired
-//    AdminRepository adminRepository;
-//    @Autowired
-//    TechStackRepository techStackRepository;
+    @Autowired
+    RestTemplate restTemplate;
 
     /*
      * Purpose : Implement the Logic of Creating Candidate Details
@@ -31,14 +31,18 @@ public class CandidateService implements ICandidateService {
      * @Param :  token ,candidateDTO and techStackId
      * */
     @Override
-    public CandidateModel addCandidate(CandidateDTO candidateDTO, Long techStackId) {
-        CandidateModel candidateModel = new CandidateModel(candidateDTO);
-        candidateModel.setRegisterDate(LocalDate.now());
-        candidateRepository.save(candidateModel);
-        String body = "Candidate Details Added With Id is : " + candidateModel.getId();
-        String subject = "Candidate Registration Successfully ...";
-        mailService.send(candidateModel.getEmail(), body, subject);
-        return candidateModel;
+    public Response addCandidate(String token, CandidateDTO candidateDTO, Long techStackId) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            CandidateModel candidateModel = new CandidateModel(candidateDTO);
+            candidateModel.setRegisterDate(LocalDate.now());
+            candidateRepository.save(candidateModel);
+            String body = "Candidate Details Added With Id is : " + candidateModel.getId();
+            String subject = "Candidate Registration Successfully ...";
+            mailService.send(candidateModel.getEmail(), body, subject);
+            return new Response(200, "Success", candidateModel);
+        }
+        throw new LMSException(400, "Token Wrong");
     }
 
     /*
@@ -47,13 +51,17 @@ public class CandidateService implements ICandidateService {
      * @Param :  token
      * */
     @Override
-    public List<CandidateModel> getAllCandidates() {
-        List<CandidateModel> isCandidates = candidateRepository.findAll();
-        if (isCandidates.size() > 0) {
-            return isCandidates;
-        } else {
-            throw new LMSException(400, "No candidates found");
+    public List<CandidateModel> getAllCandidates(String token) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            List<CandidateModel> isCandidates = candidateRepository.findAll();
+            if (isCandidates.size() > 0) {
+                return isCandidates;
+            } else {
+                throw new LMSException(400, "No candidates found");
+            }
         }
+        throw new LMSException(400, "Token Wrong");
     }
 
     /*
@@ -62,30 +70,34 @@ public class CandidateService implements ICandidateService {
      * @Param :  token and candidateDTO,id,techstackId
      * */
     @Override
-    public CandidateModel updateCandidateDetails(CandidateDTO candidateDTO, Long id, Long techStackId) {
-        Optional<CandidateModel> isCandidate = candidateRepository.findById(id);
-        if (isCandidate.isPresent()) {
-            isCandidate.get().setCicId(candidateDTO.getCicId());
-            isCandidate.get().setFullName(candidateDTO.getFullName());
-            isCandidate.get().setMobileNum(candidateDTO.getMobileNum());
-            isCandidate.get().setCandidateStatus(candidateDTO.getCandidateStatus());
-            isCandidate.get().setCity(candidateDTO.getCity());
-            isCandidate.get().setEmail(candidateDTO.getEmail());
-            isCandidate.get().setHiredDate(candidateDTO.getHiredDate());
-            isCandidate.get().setDegree(candidateDTO.getDegree());
-            isCandidate.get().setAggregatePercentage(candidateDTO.getAggregatePercentage());
-            isCandidate.get().setState(candidateDTO.getState());
-            isCandidate.get().setPassedOutYear(candidateDTO.getPassedOutYear());
-            isCandidate.get().setPreferredJobLocation(candidateDTO.getPreferredJobLocation());
-            isCandidate.get().setUpdatedDate(LocalDate.now());
-            candidateRepository.save(isCandidate.get());
-            String body = "Candidate Details Updated With Id is : " + isCandidate.get().getId();
-            String subject = "Candidate Details Updated Successfully ...";
-            mailService.send(isCandidate.get().getEmail(), body, subject);
-            return isCandidate.get();
-        } else {
-            throw new LMSException(400, "No candidates found");
+    public Response updateCandidateDetails(String token, CandidateDTO candidateDTO, Long id, Long techStackId) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Optional<CandidateModel> isCandidate = candidateRepository.findById(id);
+            if (isCandidate.isPresent()) {
+                isCandidate.get().setCicId(candidateDTO.getCicId());
+                isCandidate.get().setFullName(candidateDTO.getFullName());
+                isCandidate.get().setMobileNum(candidateDTO.getMobileNum());
+                isCandidate.get().setCandidateStatus(candidateDTO.getCandidateStatus());
+                isCandidate.get().setCity(candidateDTO.getCity());
+                isCandidate.get().setEmail(candidateDTO.getEmail());
+                isCandidate.get().setHiredDate(candidateDTO.getHiredDate());
+                isCandidate.get().setDegree(candidateDTO.getDegree());
+                isCandidate.get().setAggregatePercentage(candidateDTO.getAggregatePercentage());
+                isCandidate.get().setState(candidateDTO.getState());
+                isCandidate.get().setPassedOutYear(candidateDTO.getPassedOutYear());
+                isCandidate.get().setPreferredJobLocation(candidateDTO.getPreferredJobLocation());
+                isCandidate.get().setUpdatedDate(LocalDate.now());
+                candidateRepository.save(isCandidate.get());
+                String body = "Candidate Details Updated With Id is : " + isCandidate.get().getId();
+                String subject = "Candidate Details Updated Successfully ...";
+                mailService.send(isCandidate.get().getEmail(), body, subject);
+                return new Response(200, "Success", isCandidate.get());
+            } else {
+                throw new LMSException(400, "No candidates found");
+            }
         }
+        throw new LMSException(400, "Token Wrong");
     }
 
     /*
@@ -94,18 +106,22 @@ public class CandidateService implements ICandidateService {
      * @Param :  token and id
      * */
     @Override
-    public CandidateModel deleteCandidate(Long id) {
+    public Response deleteCandidate(String token, Long id) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
             Optional<CandidateModel> isCandidate = candidateRepository.findById(id);
             if (isCandidate.isPresent()) {
                 candidateRepository.delete(isCandidate.get());
                 String body = "Admin Details Deleted With Id is : " + isCandidate.get().getId();
                 String subject = "Admin Details Deleted Successfully ...";
                 mailService.send(isCandidate.get().getEmail(), body, subject);
-                return isCandidate.get();
+                return new Response(200, "Success", isCandidate.get());
             } else {
                 throw new LMSException(400, "No candidates found");
             }
         }
+        throw new LMSException(400, "Token Wrong");
+    }
 
     /*
      * Purpose : Implement the Logic of Get Candidate by Status
@@ -113,7 +129,9 @@ public class CandidateService implements ICandidateService {
      * @Param :  token and status
      * */
     @Override
-    public List<CandidateModel> getCandidatesByStatus(String status) {
+    public List<CandidateModel> getCandidatesByStatus(String status, String token) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
             List<CandidateModel> isCandidate = candidateRepository.findAllByStatus(status);
             if (isCandidate.size() > 0) {
                 return isCandidate;
@@ -121,6 +139,8 @@ public class CandidateService implements ICandidateService {
                 throw new LMSException(400, "No candidates found");
             }
         }
+        throw new LMSException(400, "Token Wrong");
+    }
 
     /*
      * Purpose : Implement the Logic of Update Candidate Status
@@ -128,16 +148,20 @@ public class CandidateService implements ICandidateService {
      * @Param :  token,id and status
      * */
     @Override
-    public CandidateModel changeStatus(Long id, String status) {
+    public Response changeStatus(String token, Long id, String status) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
             Optional<CandidateModel> isCandidate = candidateRepository.findById(id);
             if (isCandidate.isPresent()) {
                 isCandidate.get().setStatus(status);
                 candidateRepository.save(isCandidate.get());
-                return isCandidate.get();
+                return new Response(200, "Success", isCandidate.get());
             } else {
                 throw new LMSException(400, "No candidates found");
             }
         }
+        throw new LMSException(400, "Token Wrong");
+    }
 
     /*
      * Purpose : Implement the Logic of Get Count By Candidate Status
@@ -145,8 +169,12 @@ public class CandidateService implements ICandidateService {
      * @Param :  token and status
      * */
     @Override
-    public Long getCountByStatus(String status) {
+    public Long getCountByStatus(String token, String status) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
             Long isCandidate = candidateRepository.getCountByStatus(status);
             return isCandidate;
+        }
+        throw new LMSException(400, "Token Wrong");
     }
 }
